requirements.txt
Flask
gunicorn==23.0.0
mysql-connector-python
python-dotenv

